Add a setup wizard to simplify getting started with the plugin
The plugin settings are pretty well hidden in the WooCommerce shipping
settings, so this feature will make the plugin setup much faster as the
user is guided through the setup step-by-setp.

The setup wizard can be accessed with a page=wcpk-setup query parameter
in WP admin. Also, an admin notification with a link to the setup wizard
is shown if the plugin has been freshly installed. The "freshness" of
the installation is determined by searching the database for plugin
settings: if there are none defined, we can assume that there is a need
to run the setup wizard.

Closes: #20.